requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
comtypes==1.1.14
decorator==4.4.2
h11==0.13.0
idna==3.3
imageio==2.21.2
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.23.3
outcome==1.2.0
Pillow==9.2.0
praw==7.6.0
prawcore==2.3.0
proglog==0.1.10
pycparser==2.21
pydub==0.25.1
pypiwin32==223
PySocks==1.7.1
python-decouple==3.6
python-dotenv==0.21.0
pyttsx3==2.90
pytube==12.1.0
pywin32==304
requests==2.28.1
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
update-checker==0.18.0
urllib3==1.26.12
webdriver-manager==3.8.3
websocket-client==1.4.1
wsproto==1.2.0
